requirements.txt
streamlit
langchain
faiss-cpu
pymupdf
sentence-transformers
tqdm
ollama
langchain-community
langchain-ollama
